Create DanHansen
I've added a few changes. I now have three classes. KeyClass which is a basic list that only accepts 32 character keys. NewsAPI_KeyList which inherits the behaviour of KeyList and polymorphs the get_key function so that it tests a key and returns it if it is valid.

I have also added the three external sub to be class subs.

I'll do some more work after the quiz on Thursday.